refactor: CardGenerator 인터페이스 및 구현체 패키지 분리
