+++ src/main/java/domain/cardsGenerator/CardsGenerator.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.cardsGenerator;
 
+import domain.Card;
 import java.util.List;
 
 public interface CardsGenerator {
