+++ src/main/java/domain/cardsGenerator/RandomCardsGenerator.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.cardsGenerator;
 
+import domain.Card;
+import domain.CardNumber;
+import domain.CardShape;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
